+++ CHANGELOG.md
@@ -3,6 +3,7 @@ Version 223:
 * Add test::stream::service
 * Add websocket service
 * Pausation abandoning test
+* Destroy abandoned websocket ops on shutdown
 
 --------------------------------------------------------------------------------